organize test directory utilities into tests
