+++ tests/utils/logging.go
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: Copyright (c) 2016-2024, CloudZero, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package test
+package utils
 
 // This file provides utilities meant for use in tests that want to verify log output
